+++ ascender/Repository/IMetricRepository.cs
@@ -5,6 +5,6 @@ namespace ascender.Repository;
 public interface IMetricRepository
 {
     public void CreateMetric(CreateMetricDto dto);
-    public void MetricCommitted(string name, int value);
-    public int GetCutoff(string name);
+    public void MetricCommitted(string name, decimal value);
+    public decimal GetCutoff(string name);
 }
